[MISC] Move {core->utility}/char_operations/transform.hpp
