+++ include/seqan3/range/views/to_upper.hpp
@@ -12,10 +12,11 @@
 
 #pragma once
 
-#include <seqan3/alphabet/concept.hpp>
-#include <seqan3/core/char_operations/transform.hpp>
-#include <seqan3/range/views/deep.hpp>
 #include <seqan3/std/ranges>
+
+#include <seqan3/alphabet/concept.hpp>
+#include <seqan3/range/views/deep.hpp>
+#include <seqan3/utility/char_operations/transform.hpp>
 
 namespace seqan3::views
 {
